requirements.txt
Flask==3.1.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
